requirements.txt
pyserial
dearpygui
requests